Avoid fatal exception when authenticating
When authenticating with invalid values you will get a guzzleException, that it either cannot resolve the domain name or the login values are wrong.

Instead of letting it throw this error we set empty tokens so when why try to make another request we will get a 401 error. We can catch this error in the application that is using the package.